+++ sdk/appcontainers/arm-appcontainers/samples-dev/logicAppsListWorkflowsSample.ts
@@ -9,7 +9,7 @@ import "dotenv/config";
  * This sample demonstrates how to List the workflows for a logic app.
  *
  * @summary List the workflows for a logic app.
- * x-ms-original-file: specification/app/resource-manager/Microsoft.App/preview/2025-02-02-preview/examples/LogicApps_ListWorkflows.json
+ * x-ms-original-file: specification/app/resource-manager/Microsoft.App/ContainerApps/stable/2025-07-01/examples/LogicApps_ListWorkflows.json
  */
 async function listTheWorkflows(): Promise<void> {
   const subscriptionId =