+++ sdk/containerservice/arm-containerservice/samples/v23/typescript/src/managedClustersRotateClusterCertificatesSample.ts
@@ -1,0 +1,35 @@
+// Copyright (c) Microsoft Corporation.
+// Licensed under the MIT License.
+
+import { ContainerServiceClient } from "@azure/arm-containerservice";
+import { DefaultAzureCredential } from "@azure/identity";
+import "dotenv/config";
+
+/**
+ * This sample demonstrates how to See [Certificate rotation](https://docs.microsoft.com/azure/aks/certificate-rotation) for more details about rotating managed cluster certificates.
+ *
+ * @summary See [Certificate rotation](https://docs.microsoft.com/azure/aks/certificate-rotation) for more details about rotating managed cluster certificates.
+ * x-ms-original-file: specification/containerservice/resource-manager/Microsoft.ContainerService/aks/stable/2025-08-01/examples/ManagedClustersRotateClusterCertificates.json
+ */
+async function rotateClusterCertificates(): Promise<void> {
+  const subscriptionId =
+    process.env["CONTAINERSERVICE_SUBSCRIPTION_ID"] ||
+    "00000000-0000-0000-0000-000000000000";
+  const resourceGroupName =
+    process.env["CONTAINERSERVICE_RESOURCE_GROUP"] || "rg1";
+  const resourceName = "clustername1";
+  const credential = new DefaultAzureCredential();
+  const client = new ContainerServiceClient(credential, subscriptionId);
+  const result =
+    await client.managedClusters.beginRotateClusterCertificatesAndWait(
+      resourceGroupName,
+      resourceName,
+    );
+  console.log(result);
+}
+
+async function main(): Promise<void> {
+  await rotateClusterCertificates();
+}
+
+main().catch(console.error);