+++ sdk/netapp/arm-netapp/src/api/backupsUnderVolume/operations.ts
@@ -1,0 +1,87 @@
+// Copyright (c) Microsoft Corporation.
+// Licensed under the MIT License.
+
+import type { NetAppManagementContext as Client } from "../index.js";
+import type { BackupsMigrationRequest } from "../../models/models.js";
+import {
+  errorResponseDeserializer,
+  backupsMigrationRequestSerializer,
+} from "../../models/models.js";
+import { getLongRunningPoller } from "../../static-helpers/pollingHelpers.js";
+import { expandUrlTemplate } from "../../static-helpers/urlTemplate.js";
+import type { BackupsUnderVolumeMigrateBackupsOptionalParams } from "./options.js";
+import type { StreamableMethod, PathUncheckedResponse } from "@azure-rest/core-client";
+import { createRestError, operationOptionsToRequestParameters } from "@azure-rest/core-client";
+import type { PollerLike, OperationState } from "@azure/core-lro";
+
+export function _migrateBackupsSend(
+  context: Client,
+  resourceGroupName: string,
+  accountName: string,
+  poolName: string,
+  volumeName: string,
+  body: BackupsMigrationRequest,
+  options: BackupsUnderVolumeMigrateBackupsOptionalParams = {
+    requestOptions: {},
+  },
+): StreamableMethod {
+  const path = expandUrlTemplate(
+    "/subscriptions/{subscriptionId}/resourceGroups/{resourceGroupName}/providers/Microsoft.NetApp/netAppAccounts/{accountName}/capacityPools/{poolName}/volumes/{volumeName}/migrateBackups{?api%2Dversion}",
+    {
+      subscriptionId: context.subscriptionId,
+      resourceGroupName: resourceGroupName,
+      accountName: accountName,
+      poolName: poolName,
+      volumeName: volumeName,
+      "api%2Dversion": context.apiVersion,
+    },
+    {
+      allowReserved: options?.requestOptions?.skipUrlEncoding,
+    },
+  );
+  return context.path(path).post({
+    ...operationOptionsToRequestParameters(options),
+    contentType: "application/json",
+    body: backupsMigrationRequestSerializer(body),
+  });
+}
+
+export async function _migrateBackupsDeserialize(result: PathUncheckedResponse): Promise<void> {
+  const expectedStatuses = ["202", "200"];
+  if (!expectedStatuses.includes(result.status)) {
+    const error = createRestError(result);
+    error.details = errorResponseDeserializer(result.body);
+    throw error;
+  }
+
+  return;
+}
+
+/** Migrate the backups under volume to backup vault */
+export function migrateBackups(
+  context: Client,
+  resourceGroupName: string,
+  accountName: string,
+  poolName: string,
+  volumeName: string,
+  body: BackupsMigrationRequest,
+  options: BackupsUnderVolumeMigrateBackupsOptionalParams = {
+    requestOptions: {},
+  },
+): PollerLike<OperationState<void>, void> {
+  return getLongRunningPoller(context, _migrateBackupsDeserialize, ["202", "200"], {
+    updateIntervalInMs: options?.updateIntervalInMs,
+    abortSignal: options?.abortSignal,
+    getInitialResponse: () =>
+      _migrateBackupsSend(
+        context,
+        resourceGroupName,
+        accountName,
+        poolName,
+        volumeName,
+        body,
+        options,
+      ),
+    resourceLocationConfig: "location",
+  }) as PollerLike<OperationState<void>, void>;
+}