+++ sdk/network/arm-network/samples-dev/azureFirewallsListSample.ts
@@ -1,17 +1,16 @@
 // Copyright (c) Microsoft Corporation.
 // Licensed under the MIT License.
 
-/**
- * This sample demonstrates how to Lists all Azure Firewalls in a resource group.
- *
- * @summary Lists all Azure Firewalls in a resource group.
- * x-ms-original-file: specification/network/resource-manager/Microsoft.Network/stable/2024-07-01/examples/AzureFirewallListByResourceGroup.json
- */
-
 import { NetworkManagementClient } from "@azure/arm-network";
 import { DefaultAzureCredential } from "@azure/identity";
 import "dotenv/config";
 
+/**
+ * This sample demonstrates how to Lists all Azure Firewalls in a resource group.
+ *
+ * @summary Lists all Azure Firewalls in a resource group.
+ * x-ms-original-file: specification/network/resource-manager/Microsoft.Network/stable/2024-10-01/examples/AzureFirewallListByResourceGroup.json
+ */
 async function listAllAzureFirewallsForAGivenResourceGroup(): Promise<void> {
   const subscriptionId = process.env["NETWORK_SUBSCRIPTION_ID"] || "subid";
   const resourceGroupName = process.env["NETWORK_RESOURCE_GROUP"] || "rg1";