+++ sdk/network/arm-network/samples-dev/networkVirtualAppliancesRestartSample.ts
@@ -1,17 +1,16 @@
 // Copyright (c) Microsoft Corporation.
 // Licensed under the MIT License.
 
-/**
- * This sample demonstrates how to Restarts one or more VMs belonging to the specified Network Virtual Appliance.
- *
- * @summary Restarts one or more VMs belonging to the specified Network Virtual Appliance.
- * x-ms-original-file: specification/network/resource-manager/Microsoft.Network/stable/2024-07-01/examples/NetworkVirtualApplianceEmptyRestart.json
- */
-
 import { NetworkManagementClient } from "@azure/arm-network";
 import { DefaultAzureCredential } from "@azure/identity";
 import "dotenv/config";
 
+/**
+ * This sample demonstrates how to Restarts one or more VMs belonging to the specified Network Virtual Appliance.
+ *
+ * @summary Restarts one or more VMs belonging to the specified Network Virtual Appliance.
+ * x-ms-original-file: specification/network/resource-manager/Microsoft.Network/stable/2024-10-01/examples/NetworkVirtualApplianceEmptyRestart.json
+ */
 async function restartAllNetworkVirtualApplianceVMSInVMScaleSet(): Promise<void> {
   const subscriptionId = process.env["NETWORK_SUBSCRIPTION_ID"] || "subid";
   const resourceGroupName = process.env["NETWORK_RESOURCE_GROUP"] || "rg1";
@@ -29,7 +28,7 @@ async function restartAllNetworkVirtualApplianceVMSInVMScaleSet(): Promise<void>
  * This sample demonstrates how to Restarts one or more VMs belonging to the specified Network Virtual Appliance.
  *
  * @summary Restarts one or more VMs belonging to the specified Network Virtual Appliance.
- * x-ms-original-file: specification/network/resource-manager/Microsoft.Network/stable/2024-07-01/examples/NetworkVirtualApplianceSpecificRestart.json
+ * x-ms-original-file: specification/network/resource-manager/Microsoft.Network/stable/2024-10-01/examples/NetworkVirtualApplianceSpecificRestart.json
  */
 async function restartSpecificNetworkVirtualApplianceVMSInVMScaleSet(): Promise<void> {
   const subscriptionId = process.env["NETWORK_SUBSCRIPTION_ID"] || "subid";