+++ sdk/redisenterprise/arm-redisenterprisecache/samples-dev/databasesListKeysSample.ts
@@ -1,17 +1,16 @@
 // Copyright (c) Microsoft Corporation.
 // Licensed under the MIT License.
 
-/**
- * This sample demonstrates how to Retrieves the access keys for the Redis Enterprise database.
- *
- * @summary Retrieves the access keys for the Redis Enterprise database.
- * x-ms-original-file: specification/redisenterprise/resource-manager/Microsoft.Cache/preview/2025-05-01-preview/examples/RedisEnterpriseDatabasesListKeys.json
- */
-
 import { RedisEnterpriseManagementClient } from "@azure/arm-redisenterprisecache";
 import { DefaultAzureCredential } from "@azure/identity";
 import "dotenv/config";
 
+/**
+ * This sample demonstrates how to Retrieves the access keys for the Redis Enterprise database.
+ *
+ * @summary Retrieves the access keys for the Redis Enterprise database.
+ * x-ms-original-file: specification/redisenterprise/resource-manager/Microsoft.Cache/RedisEnterprise/stable/2025-07-01/examples/RedisEnterpriseDatabasesListKeys.json
+ */
 async function redisEnterpriseDatabasesListKeys(): Promise<void> {
   const subscriptionId =
     process.env["REDISENTERPRISE_SUBSCRIPTION_ID"] ||